Update py_ci_cd.yml workflow
This commit updates the py_ci_cd.yml workflow file. It includes changes to the keyword_check job, specifically moving the outputs section below the steps section. Additionally, it adds a new step to get the Python version from the setup.py file and sets it as an output. Finally, it updates the setup_python step to use the Python version obtained from the previous step.